python/akeyless_vault/requirements.txt
akeyless==3.2.2
python-dotenv==0.21.0
